% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNF.R
\name{getBigNetwork}
\alias{getBigNetwork}
\title{Returns the default integrated matrix as well as impacts of the original three modalities on the integrated matrix}
\usage{
getBigNetwork()
}
\value{
The default integrated matrix
}
\description{
Returns the default integrated matrix as well as impacts of the original three modalities on the integrated matrix
}
