% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNF.R
\name{getDrugInfo}
\alias{getDrugInfo}
\title{Returns information related to a drug}
\usage{
getDrugInfo(drug)
}
\arguments{
\item{drug}{The name of the drug}
}
\value{
The CHEMBL ID of the drug as well as a list of drug targets
}
\description{
Returns information related to a drug
}
