% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNF.R
\name{getDrugsOfTargets}
\alias{getDrugsOfTargets}
\title{Returns the drugs with input targets}
\usage{
getDrugsOfTargets(targets)
}
\arguments{
\item{targets}{The names of the drug targets}
}
\value{
A list of drugs
}
\description{
Returns the drugs with input targets
}
