% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNF.R
\name{getNewNetwork}
\alias{getNewNetwork}
\title{Returns the a small network}
\usage{
getNewNetwork(newData)
}
\arguments{
\item{newData}{User's own data in JSON format (converted to an R data type)}
}
\value{
The new network
}
\description{
Returns the a small network
}
