% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNF.R
\name{getNewNetwork}
\alias{getNewNetwork}
\title{Returns a network that integrates all data modalities}
\usage{
getNewNetwork(...)
}
\value{
The new network
}
\description{
Returns a network that integrates all data modalities
}
