% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNF.R
\name{getNewNetwork}
\alias{getNewNetwork}
\title{Returns a network that integrates all data modalities}
\usage{
getNewNetwork(...)
}
\arguments{
\item{...}{The three types of drug data in the order of perturbation, sensitivity, and structure}
}
\value{
The new network
}
\description{
Returns a network that integrates all data modalities
}
